Crate cumulus_pallet_aura_ext

Cumulus extension pallet for AuRa

This pallet extends the Substrate AuRa pallet to make it compatible with parachains. It provides the Pallet, the Config and the GenesisConfig.

It is also required that the parachain runtime uses the provided BlockExecutor to properly check the constructed block on the relay chain.

cumulus_pallet_parachain_system::register_validate_block! {
    Runtime = Runtime,
    BlockExecutor = cumulus_pallet_aura_ext::BlockExecutor::<Runtime, Executive>,
}

Re-exports

pub use consensus_hook::FixedVelocityConsensusHook;

pub use pallet::*;

Modules

consensus_hook

The definition of a FixedVelocityConsensusHook for consensus logic to manage block velocity.

migration

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

Structs

BlockExecutor

The block executor used when validating a PoV at the relay chain.